Create minimum dimensions of a square photo
When any photo is uploaded, the following events may occur:

If either the width or height is less than L, the user is prompted to upload another one. Print "UPLOAD ANOTHER" in this case.
If both the width and height are large enough:
(a) If the photo is already a square, then it is accepted. Print "ACCEPTED" in this case.(b) Otherwise, the user is prompted to crop it. Print "CROP IT" in this case.